Chronify_backend/requirements.txt
# Django Core
Django==5.1.4             # Django framework
djangorestframework==3.15.2  # Django REST Framework for APIs
django-filter==24.3    # Filtering support for DRF

# Authentication
django-allauth==65.3.0 # Google Authentication
djangorestframework-simplejwt==5.3.1  # JWT authentication

# Websockets and Chat
channels==4.2.0         # Django Channels for WebSockets
channels-redis==4.2.1   # Redis integration for Django Channels

# Payment Integration
razorpay==1.4.2         # Razorpay payment gateway
stripe==11.3.0         # Stripe payment gateway

# Background Tasks
celery==5.4.0          # Celery for async tasks
redis==5.2.0          # Redis for Celery broker and WebSocket support

# Notifications
django-celery-email==3.0.0 # Asynchronous email notifications
twilio==9.3.7               # SMS notifications

# Chatbots
# rasa==3.6.20             # Rasa for chatbot integration (optional, use if needed)

# Utilities
python-decouple==3.8    # Environment variable management
pillow==11.0.0         # Image processing (e.g., profile pictures, QR codes)
qrcode==8.0           # QR code generation for tickets
django-cors-headers==4.6.0  # Enable CORS for frontend-backend communication
djangorestframework-jsonapi==7.1.0  # JSON:API support for DRF

# Database
mysqlclient==2.2.6      # MySQL database support
django-environ==0.11.2  # For environment variables in settings

# Testing and Debugging
pytest-django    # Testing framework for Django
pytest           # Python testing framework
coverage         # Code coverage tool
django-debug-toolbar  # Debugging tool for development

# Deployment
gunicorn        # WSGI HTTP server for deployment
whitenoise       # Serve static files in production

